fix: remove int32 cast landmines for long indexing migration
Audited 67 locations with (int) casts that could silently overflow
for arrays > 2 billion elements. Fixed 9 critical issues:

Storage/Allocation:
- UnmanagedStorage.cs: Remove (int)shape.size casts - long overload exists
- IndexCollector.cs: Add MaxCapacity check to prevent growth beyond Array.MaxLength

Shape/Reshape:
- np.meshgrid.cs: Use reshape(long, long) overload directly
- np.nanvar.cs, np.nanstd.cs, np.nanmean.cs: Remove (int) cast in List<long>.Add

Array Conversion:
- NdArrayToMultiDimArray.cs: Add overflow check before converting to int[]

Verified 30+ locations already have proper guards (Span creations, string ops,
SIMD gather fallbacks, Shape operators with checked(), etc.)

Documented 7 known .NET limitations (Array.IndexOf, GetLength return int).

Added docs/INT32_CAST_LANDMINES.md tracking all findings.

Build: 0 errors | Tests: 3887 passed

+++ src/NumSharp.Core/Backends/Unmanaged/UnmanagedStorage.cs
@@ -1010,7 +1010,7 @@ namespace NumSharp.Backends
             if (shape.IsEmpty)
                 throw new ArgumentNullException(nameof(shape));
 
-            _Allocate(shape, ArraySlice.Allocate(dtype ?? DType, (int)shape.size, true));
+            _Allocate(shape, ArraySlice.Allocate(dtype ?? DType, shape.size, true));
         }
 
         /// <summary>
@@ -1023,7 +1023,7 @@ namespace NumSharp.Backends
             if (shape.IsEmpty)
                 throw new ArgumentNullException(nameof(shape));
 
-            _Allocate(shape, ArraySlice.Allocate(dtype ?? DType, (int)shape.size, fillZeros));
+            _Allocate(shape, ArraySlice.Allocate(dtype ?? DType, shape.size, fillZeros));
         }
 
         /// <summary>
